[sessions] Do not purge session data for foreign history deletions
This CL adds the Reason enum to history::DeletionInfo. Currently the
enum only supports kDeleteAllForeignVisits and kOther.

Foreign visits are visits that occurred on another device but were
synced to this device. More specifically they are visits with a
non-empty `originator_cache_guid`.

Foreign visits may be deleted from history in cases where history
sync was previously enabled but is now disabled. This cleanup can
occur at startup.

Ideally SessionService would remove only the entries that correspond
to these deleted foreign visit URLs. However currently there is no
easy way to remove specific URLs, instead all session service
persisted data is deleted.

This has the effect of Chrome deleting the user's previous session
data at startup, a cause of significant disruption. In this
specific case we should preserve the last session data to avoid user
data loss.

A follow up can wrap DeletionInfo::is_from_expiration_ into the
DeletionInfo::Reason enum.

Bug: 1424800, 1423163, 1454345
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4610021
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158334}

+++ chrome/browser/browsing_data/navigation_entry_remover.cc
@@ -217,6 +217,13 @@ void RemoveNavigationEntries(Profile* profile,
   DCHECK(!profile->IsOffTheRecord());
   DCHECK(!deletion_info.is_from_expiration());
 
+  // For debugging crbug.com/1424800. Remove once this has been resolved.
+  LOG(ERROR) << "RemoveNavigationEntries(), deleted_rows count = "
+             << deletion_info.deleted_rows().size()
+             << ", is_from_expiration = " << deletion_info.is_from_expiration()
+             << ", deletion_reason = "
+             << static_cast<int>(deletion_info.deletion_reason());
+
   base::flat_set<GURL> url_set;
   if (!deletion_info.time_range().IsValid())
     url_set = CreateUrlSet(deletion_info.deleted_rows());
@@ -225,7 +232,19 @@ void RemoveNavigationEntries(Profile* profile,
                              deletion_info.restrict_urls(), url_set);
   DeleteTabRestoreEntries(profile, deletion_info.time_range(),
                           deletion_info.restrict_urls(), url_set);
-  DeleteLastSessionFromSessionService(profile);
+
+  // Removal of navigation entries may occur at any point during runtime and
+  // session service data is cleared so that it can be later rebuilt without the
+  // deleted entries.
+  // However deletion of foreign visits specifically can occur during startup
+  // and clearing session service data will delete the user's previous session
+  // with no ability to rebuild/recover (see crbug.com/1424800). Foreign visits
+  // can't be part of the local session so there is no risk of retaining the
+  // session service data in this case.
+  if (deletion_info.deletion_reason() !=
+      history::DeletionInfo::Reason::kDeleteAllForeignVisits) {
+    DeleteLastSessionFromSessionService(profile);
+  }
 }
 
 }  // namespace browsing_data
